backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.1.0
numpy>=1.24.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
google.generativeai>=0.3.0
pytest>=8.0.0